File updates to pass testing

+++ ft_strcmp.c
@@ -6,7 +6,7 @@
 /*   By: sinkosi <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/05/20 16:11:22 by sinkosi           #+#    #+#             */
-/*   Updated: 2019/06/04 16:43:13 by sinkosi          ###   ########.fr       */
+/*   Updated: 2019/06/05 13:38:55 by sinkosi          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -35,7 +35,7 @@ int	ft_strcmp(const char *s1, const char *s2)
 	}
 	while ((str1[i] != '\0' && str2[i] != '\0') && control == 0)
 	{
-		control = ((str1[i]) - (str2[i]));
+		control = ((int)(str1[i] - str2[i]));
 		i++;
 	}
 	return (control);
